app/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform[langchain,reasoningengine]
cloudpickle==3.0.0
pydantic==2.7.4
requests
twilio
SpeechRecognition
soundfile
google-cloud-bigquery
google-cloud-storage
langchain-openai
langchain
google-cloud-vision
ruff
python-multipart
google-cloud-firestore
python-dotenv